Struct aws_sdk_elasticloadbalancing::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for ElasticLoadBalancing_v7.

This client allows ergonomic access to a ElasticLoadBalancing_v7-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_tags(&self) -> AddTags<C, M, R>

pub fn apply_security_groups_to_load_balancer(
&self
) -> ApplySecurityGroupsToLoadBalancer<C, M, R>

pub fn attach_load_balancer_to_subnets(
&self
) -> AttachLoadBalancerToSubnets<C, M, R>

pub fn configure_health_check(&self) -> ConfigureHealthCheck<C, M, R>

pub fn create_app_cookie_stickiness_policy(
&self
) -> CreateAppCookieStickinessPolicy<C, M, R>

pub fn create_lb_cookie_stickiness_policy(
&self
) -> CreateLBCookieStickinessPolicy<C, M, R>

pub fn create_load_balancer(&self) -> CreateLoadBalancer<C, M, R>

pub fn create_load_balancer_listeners(
&self
) -> CreateLoadBalancerListeners<C, M, R>

pub fn create_load_balancer_policy(&self) -> CreateLoadBalancerPolicy<C, M, R>

pub fn delete_load_balancer(&self) -> DeleteLoadBalancer<C, M, R>

pub fn delete_load_balancer_listeners(
&self
) -> DeleteLoadBalancerListeners<C, M, R>

pub fn delete_load_balancer_policy(&self) -> DeleteLoadBalancerPolicy<C, M, R>

pub fn deregister_instances_from_load_balancer(
&self
) -> DeregisterInstancesFromLoadBalancer<C, M, R>

pub fn describe_account_limits(&self) -> DescribeAccountLimits<C, M, R>

pub fn describe_instance_health(&self) -> DescribeInstanceHealth<C, M, R>

pub fn describe_load_balancer_attributes(
&self
) -> DescribeLoadBalancerAttributes<C, M, R>

pub fn describe_load_balancer_policies(
&self
) -> DescribeLoadBalancerPolicies<C, M, R>

pub fn describe_load_balancer_policy_types(
&self
) -> DescribeLoadBalancerPolicyTypes<C, M, R>

pub fn describe_load_balancers(&self) -> DescribeLoadBalancers<C, M, R>

pub fn describe_tags(&self) -> DescribeTags<C, M, R>

pub fn detach_load_balancer_from_subnets(
&self
) -> DetachLoadBalancerFromSubnets<C, M, R>

pub fn disable_availability_zones_for_load_balancer(
&self
) -> DisableAvailabilityZonesForLoadBalancer<C, M, R>

pub fn enable_availability_zones_for_load_balancer(
&self
) -> EnableAvailabilityZonesForLoadBalancer<C, M, R>

pub fn modify_load_balancer_attributes(
&self
) -> ModifyLoadBalancerAttributes<C, M, R>

pub fn register_instances_with_load_balancer(
&self
) -> RegisterInstancesWithLoadBalancer<C, M, R>

pub fn remove_tags(&self) -> RemoveTags<C, M, R>

pub fn set_load_balancer_listener_ssl_certificate(
&self
) -> SetLoadBalancerListenerSSLCertificate<C, M, R>

pub fn set_load_balancer_policies_for_backend_server(
&self
) -> SetLoadBalancerPoliciesForBackendServer<C, M, R>

pub fn set_load_balancer_policies_of_listener(
&self
) -> SetLoadBalancerPoliciesOfListener<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self