Module aws_sdk_elasticloadbalancing::error

Modules

access_point_not_found_exception

See AccessPointNotFoundException

certificate_not_found_exception

See CertificateNotFoundException

dependency_throttle_exception

See DependencyThrottleException

duplicate_access_point_name_exception

See DuplicateAccessPointNameException

duplicate_listener_exception

See DuplicateListenerException

duplicate_policy_name_exception

See DuplicatePolicyNameException

duplicate_tag_keys_exception

See DuplicateTagKeysException

invalid_configuration_request_exception

See InvalidConfigurationRequestException

invalid_end_point_exception

See InvalidEndPointException

invalid_scheme_exception

See InvalidSchemeException

invalid_security_group_exception

See InvalidSecurityGroupException

invalid_subnet_exception

See InvalidSubnetException

listener_not_found_exception

See ListenerNotFoundException

load_balancer_attribute_not_found_exception

See LoadBalancerAttributeNotFoundException

operation_not_permitted_exception

See OperationNotPermittedException

policy_not_found_exception

See PolicyNotFoundException

policy_type_not_found_exception

See PolicyTypeNotFoundException

subnet_not_found_exception

See SubnetNotFoundException

too_many_access_points_exception

See TooManyAccessPointsException

too_many_policies_exception

See TooManyPoliciesException

too_many_tags_exception

See TooManyTagsException

unsupported_protocol_exception

See UnsupportedProtocolException

Structs

AccessPointNotFoundException

The specified load balancer does not exist.

AddTagsError

ApplySecurityGroupsToLoadBalancerError

AttachLoadBalancerToSubnetsError

CertificateNotFoundException

The specified ARN does not refer to a valid SSL certificate in AWS Identity and Access Management (IAM)

ConfigureHealthCheckError

CreateAppCookieStickinessPolicyError

CreateLBCookieStickinessPolicyError

CreateLoadBalancerError

CreateLoadBalancerListenersError

CreateLoadBalancerPolicyError

DeleteLoadBalancerError

DeleteLoadBalancerListenersError

DeleteLoadBalancerPolicyError

DependencyThrottleException

A request made by Elastic Load Balancing to another service exceeds the maximum request rate permitted for your account.

DeregisterInstancesFromLoadBalancerError

DescribeAccountLimitsError

DescribeInstanceHealthError

DescribeLoadBalancerAttributesError

DescribeLoadBalancerPoliciesError

DescribeLoadBalancerPolicyTypesError

DescribeLoadBalancersError

DescribeTagsError

DetachLoadBalancerFromSubnetsError

DisableAvailabilityZonesForLoadBalancerError

DuplicateAccessPointNameException

The specified load balancer name already exists for this account.

DuplicateListenerException

A listener already exists for the specified load balancer name and port, but with a different instance port, protocol, or SSL certificate.

DuplicatePolicyNameException

A policy with the specified name already exists for this load balancer.

DuplicateTagKeysException

A tag key was specified more than once.

EnableAvailabilityZonesForLoadBalancerError

InvalidConfigurationRequestException

The requested configuration change is not valid.

InvalidEndPointException

The specified endpoint is not valid.

InvalidSchemeException

The specified value for the schema is not valid. You can only specify a scheme for load balancers in a VPC.

InvalidSecurityGroupException

One or more of the specified security groups do not exist.

InvalidSubnetException

The specified VPC has no associated Internet gateway.

ListenerNotFoundException

The load balancer does not have a listener configured at the specified port.

LoadBalancerAttributeNotFoundException

The specified load balancer attribute does not exist.

ModifyLoadBalancerAttributesError

OperationNotPermittedException

This operation is not allowed.

PolicyNotFoundException

One or more of the specified policies do not exist.

PolicyTypeNotFoundException

One or more of the specified policy types do not exist.

RegisterInstancesWithLoadBalancerError

RemoveTagsError

SetLoadBalancerListenerSSLCertificateError

SetLoadBalancerPoliciesForBackendServerError

SetLoadBalancerPoliciesOfListenerError

SubnetNotFoundException

One or more of the specified subnets do not exist.

TooManyAccessPointsException

The quota for the number of load balancers has been reached.

TooManyPoliciesException

The quota for the number of policies for this load balancer has been reached.

TooManyTagsException

The quota for the number of tags that can be assigned to a load balancer has been reached.

UnsupportedProtocolException

The specified protocol or signature version is not supported.

Enums

AddTagsErrorKind

ApplySecurityGroupsToLoadBalancerErrorKind

AttachLoadBalancerToSubnetsErrorKind

ConfigureHealthCheckErrorKind

CreateAppCookieStickinessPolicyErrorKind

CreateLBCookieStickinessPolicyErrorKind

CreateLoadBalancerErrorKind

CreateLoadBalancerListenersErrorKind

CreateLoadBalancerPolicyErrorKind

DeleteLoadBalancerErrorKind

DeleteLoadBalancerListenersErrorKind

DeleteLoadBalancerPolicyErrorKind

DeregisterInstancesFromLoadBalancerErrorKind

DescribeAccountLimitsErrorKind

DescribeInstanceHealthErrorKind

DescribeLoadBalancerAttributesErrorKind

DescribeLoadBalancerPoliciesErrorKind

DescribeLoadBalancerPolicyTypesErrorKind

DescribeLoadBalancersErrorKind

DescribeTagsErrorKind

DetachLoadBalancerFromSubnetsErrorKind

DisableAvailabilityZonesForLoadBalancerErrorKind

EnableAvailabilityZonesForLoadBalancerErrorKind

ModifyLoadBalancerAttributesErrorKind

RegisterInstancesWithLoadBalancerErrorKind

RemoveTagsErrorKind

SetLoadBalancerListenerSSLCertificateErrorKind

SetLoadBalancerPoliciesForBackendServerErrorKind

SetLoadBalancerPoliciesOfListenerErrorKind