Module aws_sdk_elasticloadbalancing::operation

Structs

AddTags

Adds the specified tags to the specified load balancer. Each load balancer can have a maximum of 10 tags.

ApplySecurityGroupsToLoadBalancer

Associates one or more security groups with your load balancer in a virtual private cloud (VPC). The specified security groups override the previously associated security groups.

AttachLoadBalancerToSubnets

Adds one or more subnets to the set of configured subnets for the specified load balancer.

ConfigureHealthCheck

Specifies the health check settings to use when evaluating the health state of your EC2 instances.

CreateAppCookieStickinessPolicy

Generates a stickiness policy with sticky session lifetimes that follow that of an application-generated cookie. This policy can be associated only with HTTP/HTTPS listeners.

CreateLBCookieStickinessPolicy

Generates a stickiness policy with sticky session lifetimes controlled by the lifetime of the browser (user-agent) or a specified expiration period. This policy can be associated only with HTTP/HTTPS listeners.

CreateLoadBalancer

Creates a Classic Load Balancer.

CreateLoadBalancerListeners

Creates one or more listeners for the specified load balancer. If a listener with the specified port does not already exist, it is created; otherwise, the properties of the new listener must match the properties of the existing listener.

CreateLoadBalancerPolicy

Creates a policy with the specified attributes for the specified load balancer.

DeleteLoadBalancer

Deletes the specified load balancer.

DeleteLoadBalancerListeners

Deletes the specified listeners from the specified load balancer.

DeleteLoadBalancerPolicy

Deletes the specified policy from the specified load balancer. This policy must not be enabled for any listeners.

DeregisterInstancesFromLoadBalancer

Deregisters the specified instances from the specified load balancer. After the instance is deregistered, it no longer receives traffic from the load balancer.

DescribeAccountLimits

Describes the current Elastic Load Balancing resource limits for your AWS account.

DescribeInstanceHealth

Describes the state of the specified instances with respect to the specified load balancer. If no instances are specified, the call describes the state of all instances that are currently registered with the load balancer. If instances are specified, their state is returned even if they are no longer registered with the load balancer. The state of terminated instances is not returned.

DescribeLoadBalancerAttributes

Describes the attributes for the specified load balancer.

DescribeLoadBalancerPolicies

Describes the specified policies.

DescribeLoadBalancerPolicyTypes

Describes the specified load balancer policy types or all load balancer policy types.

DescribeLoadBalancers

Describes the specified the load balancers. If no load balancers are specified, the call describes all of your load balancers.

DescribeTags

Describes the tags associated with the specified load balancers.

DetachLoadBalancerFromSubnets

Removes the specified subnets from the set of configured subnets for the load balancer.

DisableAvailabilityZonesForLoadBalancer

Removes the specified Availability Zones from the set of Availability Zones for the specified load balancer

EnableAvailabilityZonesForLoadBalancer

Adds the specified Availability Zones to the set of Availability Zones for the specified load balancer

ModifyLoadBalancerAttributes

Modifies the attributes of the specified load balancer.

RegisterInstancesWithLoadBalancer

Adds the specified instances to the specified load balancer.

RemoveTags

Removes one or more tags from the specified load balancer.

SetLoadBalancerListenerSSLCertificate

Sets the certificate that terminates the specified listener's SSL connections. The specified certificate replaces any prior certificate that was used on the same load balancer and port.

SetLoadBalancerPoliciesForBackendServer

Replaces the set of policies associated with the specified port on which the EC2 instance is listening with a new set of policies.

SetLoadBalancerPoliciesOfListener

Replaces the current set of policies for the specified load balancer port with the specified set of policies.